requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
crispy-bootstrap4==2024.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.1
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-heroku==0.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
filelock==3.16.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
qrcode==8.0
requests==2.32.3
sqlparse==0.5.1
stripe==11.6.0
typing_extensions==4.13.0
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.27.0
whitenoise==6.9.0
